backend/app/requirements.txt
flask
flask-cors
werkzeug
transformers>=4.49
torch
torchvision
opencv-python
pillow
pdf2image
huggingface-hub
PyPDF2
poppler-utils